issue. 크루 데이터 crew_id에 조건에 맞게 데이터 조회 이슈 해결
